chore(refactor): move all scripts, generators to `hack` directory
