+++ hack/generators/cache-stores/spec.go
