+++ hack/generators/cache-stores/templates.go
