+++ ingress-controller/internal/store/zz_generated.cache_stores_test.go
@@ -5,18 +5,18 @@ package store_test
 import (
 	"testing"
 
-	"github.com/kong/kong-operator/ingress-controller/internal/gatewayapi"
 	kongv1 "github.com/kong/kong-operator/api/configuration/v1"
 	kongv1alpha1 "github.com/kong/kong-operator/api/configuration/v1alpha1"
+	"github.com/kong/kong-operator/ingress-controller/internal/gatewayapi"
 	"github.com/stretchr/testify/require"
 	corev1 "k8s.io/api/core/v1"
 	discoveryv1 "k8s.io/api/discovery/v1"
 	netv1 "k8s.io/api/networking/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/kong/kong-operator/ingress-controller/internal/store"
 	kongv1beta1 "github.com/kong/kong-operator/api/configuration/v1beta1"
 	incubatorv1alpha1 "github.com/kong/kong-operator/api/incubator/v1alpha1"
+	"github.com/kong/kong-operator/ingress-controller/internal/store"
 )
 
 func TestCacheStores(t *testing.T) {
@@ -24,6 +24,7 @@ func TestCacheStores(t *testing.T) {
 		name          string
 		objectToStore client.Object
 	}{
+
 		{
 			name:          "Ingress",
 			objectToStore: &netv1.Ingress{},
